Revamp of array views
This PR changes the definition of array views to not require a bijection
into the full array, nor full ownership of it. This allows to have
views that work only over portions of a base array, which is required
for tiling matrices. See Kuiper.ArrayView.Test.EvenOdds for one basic
(though verbose) example there.

The generated code is now polluted with some trivial pointer casts. That
is unfortunate and I don't know where it's coming from, but I don't
think it's a blocker.