Remove the dangerous SZ.t ~> erased nat coercions
These kick in when we do not want them to (e.g. if the expected type is
erased bool). Just kill them. Things look more predictable now, but this
may just be a first impression. I also changed `erased nat` to `nat` in
many types and ghost functions, which is the right thing to do, and now
it does in fact seem to work.