KernelDesc: bake in barrier contract instead of allowing creation
This gets rid of the `can_create_barrier` and
`consume_can_create_barrier` which are used to restrict barrier creation
to the `block_setup` phase.  CUDA (block-wide) barriers are very
intrinsic to kernels, so it makes sense to bake this into the kernel
description, where one can of course decide to use a trivial empty
barrier if there is no actual use for one.

This also vectorizes the shmem loading phase of TensorCore (non 2D), as
I was touching that code anyway.